refactor(dom-contract): share note story block helpers
Move `getNoteStoryKind`, `isNonBodyStoryBlockId`, and the `NoteStoryKind`
type from `painter-dom` into `@superdoc/dom-contract` so consumers like
`super-editor`'s `DomPositionIndex` can depend on the contract package
directly instead of pulling in `painter-dom`. `painter-dom` re-exports
the symbols to keep its public API stable.
