+++ packages/layout-engine/dom-contract/src/index.test.ts
@@ -15,6 +15,8 @@ import {
   DRAGGABLE_SELECTOR,
   encodeLayoutStoryDataset,
   decodeLayoutStoryDataset,
+  getNoteStoryKind,
+  isNonBodyStoryBlockId,
 } from './index.js';
 
 describe('@superdoc/dom-contract', () => {
@@ -129,4 +131,22 @@ describe('@superdoc/dom-contract', () => {
   it('exports DRAGGABLE_SELECTOR constant', () => {
     expect(DRAGGABLE_SELECTOR).toBe('[data-draggable="true"]');
   });
+
+  it.each([
+    ['footnote-1-abc', 'footnote'],
+    ['endnote-1-abc', 'endnote'],
+    ['__sd_semantic_footnote-1-abc', 'semantic-footnote'],
+    ['__sd_semantic_endnote-1-abc', 'semantic-endnote'],
+  ] as const)('detects %s as a non-body %s story block', (blockId, kind) => {
+    expect(getNoteStoryKind(blockId)).toBe(kind);
+    expect(isNonBodyStoryBlockId(blockId)).toBe(true);
+  });
+
+  it.each(['body-paragraph-1', 'footnotes-heading', '__sd_semantic_footnotes_heading', undefined])(
+    'does not treat %s as a note body fragment',
+    (blockId) => {
+      expect(getNoteStoryKind(blockId)).toBeUndefined();
+      expect(isNonBodyStoryBlockId(blockId)).toBe(false);
+    },
+  );
 });
