+++ packages/layout-engine/painters/dom/src/notes/story.ts
@@ -1,27 +1,7 @@
-export type NoteStoryKind = 'footnote' | 'endnote' | 'semantic-footnote' | 'semantic-endnote';
+import { getNoteStoryKind } from '@superdoc/dom-contract';
 
-export const getNoteStoryKind = (blockId: string | undefined): NoteStoryKind | undefined => {
-  if (typeof blockId !== 'string') {
-    return undefined;
-  }
-
-  if (blockId.startsWith('footnote-')) {
-    return 'footnote';
-  }
-  if (blockId.startsWith('endnote-')) {
-    return 'endnote';
-  }
-  if (blockId.startsWith('__sd_semantic_footnote-')) {
-    return 'semantic-footnote';
-  }
-  if (blockId.startsWith('__sd_semantic_endnote-')) {
-    return 'semantic-endnote';
-  }
-
-  return undefined;
-};
-
-export const isNonBodyStoryBlockId = (blockId: string | undefined): boolean => getNoteStoryKind(blockId) !== undefined;
+export { getNoteStoryKind, isNonBodyStoryBlockId } from '@superdoc/dom-contract';
+export type { NoteStoryKind } from '@superdoc/dom-contract';
 
 // AIDEV-NOTE: FootnotesBuilder emits `footnote-{id}-` blocks into the body painter.
 // Endnote and semantic note blocks have dedicated editing sessions, so only plain
