+++ packages/super-editor/src/editors/v1/dom-observer/DomPositionIndex.ts
@@ -1,5 +1,4 @@
-import { DOM_CLASS_NAMES } from '@superdoc/dom-contract';
-import { isNonBodyStoryBlockId } from '@superdoc/painter-dom';
+import { DOM_CLASS_NAMES, isNonBodyStoryBlockId } from '@superdoc/dom-contract';
 import { sortedIndexBy } from 'lodash';
 import { debugLog, getSelectionDebugConfig } from '../core/presentation-editor/selection/SelectionDebug.js';
